switch jinja2.Markup import to the copy from Markupsafe
In 2013, jinja2 switched to an external copy of Markupsafe and removed
its bundled implementation. Even before then, it was a fallback for the
external copy.

jinja2 3.x started raising a deprecation warning if you relied on
`jinja2` internally importing `Markup` and exposing it, and finally
removed it entirely in 3.1.0; as of today, entering the html writer code
raises a traceback and aborts.

Explicitly depend on Markupsafe (it was always a transitive requirement
through jinja2), and import directly from it.